+++ TwitterClone/HeaderView.swift
@@ -1,0 +1,20 @@
+//
+//  HeaderView.swift
+//  TwitterClone
+//
+//  Created by Muhd Mirza on 22/8/22.
+//
+
+import SwiftUI
+
+struct HeaderView: View {
+    var body: some View {
+        Image("charleyrivers").resizable().frame(height: 200)
+    }
+}
+
+struct HeaderView_Previews: PreviewProvider {
+    static var previews: some View {
+        HeaderView()
+    }
+}